requirements.txt
pyperclip
pillow
baidu-aip
pynput